# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salus API",
        "description": "API documentation for the Salus health assistant application.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Application server"
        }
    ],
    "paths": {
...
        "/api/doctors": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "List all doctors",
                "operationId": "listDoctors",
                "responses": {
                    "200": {
                        "description": "Doctors retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Dr. Youssef Amrani"
                                                    },
                                                    "specialty": {
                                                        "type": "string",
                                                        "example": "Cardiology"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Casablanca"
                                                    },
                                                    "yearsofexperience": {
                                                        "type": "integer",
                                                        "example": 12
                                                    },
                                                    "consultation_price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 350
                                                    },
                                                    "available_days": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "Monday"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/doctors/{id}": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Get a doctor by ID",
                "operationId": "showDoctor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Doctor ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Dr. Youssef Amrani"
                                                },
                                                "specialty": {
                                                    "type": "string",
                                                    "example": "Cardiology"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "Casablanca"
                                                },
                                                "yearsofexperience": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "consultation_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 350
                                                },
                                                "available_days": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Monday"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Doctor not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Doctor not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/doctors/search": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Search doctors by specialty or city",
                "operationId": "searchDoctors",
                "parameters": [
                    {
                        "name": "specialty",
                        "in": "query",
                        "description": "Filter by specialty",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Cardiology"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filter by city",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Casablanca"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctors search completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Dr. Youssef Amrani"
                                                    },
                                                    "specialty": {
                                                        "type": "string",
                                                        "example": "Cardiology"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Casablanca"
                                                    },
                                                    "yearsofexperience": {
                                                        "type": "integer",
                                                        "example": 12
                                                    },
                                                    "consultation_price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 350
                                                    },
                                                    "available_days": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "Monday"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use a Sanctum bearer token. Example: Bearer 1|abcdef...",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}